+++ lib/AST/CMakeLists.txt
@@ -71,6 +71,7 @@ add_swift_host_library(swiftAST STATIC
   NameLookupRequests.cpp
   OperatorNameLookup.cpp
   PackConformance.cpp
+  PackExpansionMatcher.cpp
   Parameter.cpp
   Pattern.cpp
   PlatformKind.cpp
